requirements.txt
APScheduler==3.6.3
astroid==2.4.2
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pylint==2.6.0
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
wcwidth==0.2.5
wrapt==1.12.1
